feat: Introduce new UI for the update screen
- Implemented a new card-based UI for the app update screen, replacing the previous basic layout.
- The new `UpdateInfoCard` displays the app icon, title, download size, and a description.
- Added a `DownloadInfoRow` composable to show download progress, including a progress bar and a cancel icon.
- The main button now dynamically changes its text from "UPDATE" to "INSTALL" after the download is complete.
- When a download is in progress, the "UPDATE" button is replaced by the `DownloadInfoRow` to show the current status.
- Added `bytesRemaining` property to `DownloadApkModel` for future use.
- Corrected the method call in `CoreReaderFragment` from `attemptToShowDonationPopup` to `attemptToShowUpdatePopup`.
- Renamed the method in `UpdateDialogHandler` to `attemptToShowUpdatePopup` for clarity.
- Removed the unused `Composable1.kt` file.
