+++ core/src/main/java/org/kiwix/kiwixmobile/core/main/reader/CoreReaderFragment.kt
@@ -1955,7 +1955,7 @@ abstract class CoreReaderFragment :
     }
     lifecycleScope.launch {
       donationDialogHandler?.attemptToShowDonationPopup()
-      updateDialogHandler?.attemptToShowDonationPopup()
+      updateDialogHandler?.attemptToShowUpdatePopup()
     }
   }
